+++ demo/src/__tests__/App.spec.tsx
@@ -22,7 +22,7 @@ describe("App", () => {
         expect(console.error).not.toHaveBeenCalled();
     });
 
-    it("renders ErrorBoundary fallback element when draw function throws error", () => {
+    it("renders ErrorBoundary fallback element when draw function throws error", async () => {
         const spy = jest
             .spyOn(global.console, "error")
             .mockImplementation(() => {});
@@ -32,30 +32,35 @@ describe("App", () => {
         const invalidJSON = JSON.stringify([{ type: 123 }]);
         fireEvent.change(input, { target: { value: invalidJSON } });
 
-        const button = screen.getByTestId("input-submit-button");
-        fireEvent.click(button);
-
-        const errorBoundary = screen.getByTestId("failure-banner");
-        expect(errorBoundary.textContent).toEqual(
-            expect.stringMatching(
-                /^✖ Invalid discriminator value\..*\n\s*→ at type$/
-            )
+        await waitFor(
+            () => {
+                const errorBoundary = screen.getByTestId("failure-banner");
+                expect(errorBoundary.textContent).toEqual(
+                    expect.stringMatching(
+                        /^✖ Invalid discriminator value\..*\n\s*→ at type$/
+                    )
+                );
+            },
+            { timeout: 2000 }
         );
         spy.mockRestore();
     });
 
-    it("calls console error, renders Alert banner, and renders disabled download button when the input is not valid JSON", () => {
+    it("calls console error, renders Alert banner, and renders disabled download button when the input is not valid JSON", async () => {
         const consoleErrorSpy = jest
             .spyOn(console, "error")
             .mockImplementation();
 
         const input = screen.getByLabelText("Enter memory model JSON here");
         fireEvent.change(input, { target: { value: "*&#*#(@(!(" } });
-        const button = screen.getByTestId("input-submit-button");
-        fireEvent.click(button);
 
-        expect(consoleErrorSpy).toHaveBeenCalledWith(
-            expect.stringMatching(/^Error parsing inputted JSON: /)
+        await waitFor(
+            () => {
+                expect(consoleErrorSpy).toHaveBeenCalledWith(
+                    expect.stringMatching(/^Error parsing inputted JSON: /)
+                );
+            },
+            { timeout: 2000 }
         );
         const alertBanner = screen.getByTestId("failure-banner");
         expect(alertBanner.textContent).toEqual(
@@ -78,36 +83,32 @@ describe("App", () => {
         const invalidJSON = JSON.stringify([{ type: 123 }]);
         fireEvent.change(input, { target: { value: invalidJSON } });
 
-        const button = screen.getByTestId("input-submit-button");
-        fireEvent.click(button);
-
-        const errorBoundary = screen.getByTestId("failure-banner");
-        expect(errorBoundary.textContent).toEqual(
-            expect.stringMatching(
-                /^✖ Invalid discriminator value\..*\n\s*→ at type$/
-            )
+        await waitFor(
+            () => {
+                const errorBoundary = screen.getByTestId("failure-banner");
+                expect(errorBoundary.textContent).toEqual(
+                    expect.stringMatching(
+                        /^✖ Invalid discriminator value\..*\n\s*→ at type$/
+                    )
+                );
+            },
+            { timeout: 2000 }
         );
 
-        // Next, reset and input valid JSON that's also valid memory-viz JSON
-        const validJSON = JSON.stringify([
-            {
-                type: ".frame",
-                name: "__main__",
-                value: { lst1: 82, lst2: 84, p: 99, d: 10, t: 11 },
-            },
-            { type: "int", id: 82 },
-            { type: "int", id: 84 },
-            { type: "int", id: 99 },
-            { type: "int", id: 10 },
-            { type: "int", id: 11 },
-        ]);
-        fireEvent.change(input, { target: { value: validJSON } });
-        fireEvent.click(button);
+        // Next, reset by selecting a sample input that's valid memory-viz JSON.
+        fireEvent.click(screen.getByText("Sample Inputs"));
+        fireEvent.click(screen.getByText("Manual Layout"));
 
         // Finally, test that the ErrorBoundary message doesn't appear, and that the canvas is reappearing after resetting
-        await waitFor(() => {
-            expect(screen.queryByTestId("failure-banner")).toBeNull();
-        });
+        await waitFor(
+            () => {
+                expect(screen.queryByTestId("failure-banner")).toBeNull();
+            },
+            { timeout: 2000 }
+        );
+        expect((input as HTMLTextAreaElement).value).toEqual(
+            expect.stringContaining("David is cool!")
+        );
         expect(screen.getByTestId("memory-models-svg")).toBeTruthy();
         spy.mockRestore();
     });